Handle getId() returning an Integer rather than int.
git-svn-id: http://10.43.0.1/svn/lex/trunk@180 7faadb50-d407-0410-ad2f-ba9894573cc4

+++ test/com/qwirx/lex/EmdrosDatabaseTest.java
@@ -235,7 +235,7 @@ public class EmdrosDatabaseTest extends TestCase
 				ch.setConstant(cv.getName(), cv.getNewValue());
 		}
 		ch.execute();
-		return ch.getId();
+		return ch.getId().intValue();
 	}
 	
 	private int getFirstChangedRowLogId(int logId)
